Decompose CanvasRenderer.render() into 6 discrete RenderPass objects (#978)
Extract the 1062-line CanvasRenderer into a pipeline of render passes:
MaterialFlowPass, InfoLinkPass, CausalLinkPass, LoopHighlightPass,
ElementPass, and InteractionOverlayPass. Each implements the new
RenderPass interface. CanvasRenderer is now a slim ~120-line orchestrator
that iterates the pass list. Shared helpers (isHiddenAux,
shouldDimConnection) live as static methods on the RenderPass interface.
